Store synonyms for each recognised POS for every word
Previously we'd lookup wordnet and only generate inflected synonyms from
the lemma if nothing was found. When finding inflected synonyms, we'd
only use the most likely (frequent) POS of the word.

In this commit all available synonyms are found and stored in the
lexicon for each word, including those that exist in wordnet in their
orginal form (present tense verbs, singular nouns, adjectives) and each
of the recognised parts of speech from the POS frequencies data.

Note: the POS frequencies data looks to be fairly lacking. Doesn't
contain, for example, "defences".